Release pending API changes
