+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20240920/ApiModels/PersonInfo.cs
@@ -1,4 +1,4 @@
-namespace TeachingRecordSystem.Api.V3.VNext.ApiModels;
+namespace TeachingRecordSystem.Api.V3.V20240920.ApiModels;
 
 public record PersonInfo
 {
