+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20240920/Controllers/PersonsController.cs
@@ -4,10 +4,10 @@ using Swashbuckle.AspNetCore.Annotations;
 using TeachingRecordSystem.Api.Infrastructure.ModelBinding;
 using TeachingRecordSystem.Api.Infrastructure.Security;
 using TeachingRecordSystem.Api.V3.Core.Operations;
-using TeachingRecordSystem.Api.V3.VNext.Requests;
-using TeachingRecordSystem.Api.V3.VNext.Responses;
+using TeachingRecordSystem.Api.V3.V20240920.Requests;
+using TeachingRecordSystem.Api.V3.V20240920.Responses;
 
-namespace TeachingRecordSystem.Api.V3.VNext.Controllers;
+namespace TeachingRecordSystem.Api.V3.V20240920.Controllers;
 
 [Route("persons")]
 public class PersonsController(IMapper mapper) : ControllerBase
@@ -108,7 +108,7 @@ public class PersonsController(IMapper mapper) : ControllerBase
         Summary = "Mark person as deceased",
         Description = "Marks a person as deceased.")]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    [Authorize(Policy = AuthorizationPolicies.ApiKey, Roles = $"{ApiRoles.UpdatePerson}")]
+    [Authorize(Policy = AuthorizationPolicies.ApiKey, Roles = ApiRoles.UpdatePerson)]
     public async Task<IActionResult> Deceased(
         [FromRoute] string trn,
         [FromBody] SetDeceasedRequest request,
