+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20240920/Requests/FindPersonRequest.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 
-namespace TeachingRecordSystem.Api.V3.VNext.Requests;
+namespace TeachingRecordSystem.Api.V3.V20240920.Requests;
 
 public record FindPersonRequest
 {
