--- TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/VNext/Responses/AlertResponse.cs
@@ -1,8 +1,0 @@
-using TeachingRecordSystem.Api.V3.VNext.ApiModels;
-
-namespace TeachingRecordSystem.Api.V3.VNext.Responses;
-
-public record AlertResponse : Alert
-{
-    public required PersonInfo Person { get; init; }
-}
