+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/VNext/WebHookMessages/AlertCreatedNotification.cs
@@ -1,4 +1,4 @@
-using TeachingRecordSystem.Api.V3.VNext.ApiModels;
+using TeachingRecordSystem.Api.V3.V20240920.ApiModels;
 
 namespace TeachingRecordSystem.Api.V3.VNext.WebHookMessages;
