+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/VersionRegistry.cs
@@ -16,6 +16,7 @@ public static class VersionRegistry
         V3MinorVersions.V20240606,
         V3MinorVersions.V20240814,
         V3MinorVersions.V20240912,
+        V3MinorVersions.V20240920,
         V3MinorVersions.VNext,
     ];
 
@@ -58,6 +59,7 @@ public static class VersionRegistry
         public const string V20240606 = "20240606";
         public const string V20240814 = "20240814";
         public const string V20240912 = "20240912";
+        public const string V20240920 = "20240920";
         public const string VNext = VNextVersion;
     }
 }
